Use gnome-terminal-transparency instead of kitty
Turns out it has a smaller memory footprint, but it does not support
ligatures. This is a minor inconveince, however -- using only 8 MiB per
new terminal instance is a good enough thing.

- Konsole is heavy (starts from 180M) and font substitution is broken
- kitty is heavy (100M each)
- qterminal is qt5 and ignores breeze
- zutty has broken font substitution
- alacritty does not have ligatures
- wezterm is heavy (approx 100M binary)
- contour has broken rendering for git-delta
- xterm is unusable
- urxvt is unusable
- mate-terminal has no font chooser and unwanted dep
- xfce4-terminal has unwanted dep
- sakura is broken, idk where
- gnome-terminal has no transparency setting
- gnome-console uses heavily styled libadwaita and featureless
- (libvte)-based terminals lack ligature support

This is ridiculous. Maybe having ligatures is not that important, but
still-- what terminal except agetty is lightweight enough?